requirements.txt
streamlit>=1.25.0
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.21.0
matplotlib>=3.5.0
Pillow>=8.3.0
